% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_vectorized.R
\name{apply_vectorized}
\alias{apply_vectorized}
\title{Vectorize a Function Using `sapply`}
\usage{
apply_vectorized(x)
}
\arguments{
\item{x}{A numeric vector to which the `non_vectorized_function` will be applied.}
}
\value{
A numeric vector where each element is the result of applying `non_vectorized_function` to the corresponding element of `x`.
}
\description{
This function vectorizes a non-vectorized function by applying it to each element of a numeric vector `x` using the `sapply` function. It is important that `non_vectorized_function` is defined and takes a single numeric value as input.
}
\examples{
# Assuming non_vectorized_function is defined and takes a single numeric value
test_vector <- 1:10
result <- apply_vectorized(test_vector)
print(result)

}
