% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_f.R
\name{evaluate_f}
\alias{evaluate_f}
\title{Evaluate and Classify the Output of a Vectorized Function}
\usage{
evaluate_f(x)
}
\arguments{
\item{x}{A numeric vector to which the `logical_index_vectorized` function will be applied.}
}
\value{
A numeric vector of the same length as `x`, where each element is the result of applying `logical_index_vectorized` to the corresponding element of `x`. The returned vector has a class attribute set to 'evaluation_of_f'.
}
\description{
Evaluate and Classify the Output of a Vectorized Function
}
\examples{
test_vector <- c(-1, 0.25, 5, 12, 16, 21)
result <- evaluate_f(test_vector)
print(result)
class(result) # Should return 'evaluation_of_f'

}
\author{
Samy Maurer

This function evaluates a vectorized version of a piecewise function on a numeric vector `x` and then sets a specific class, 'evaluation_of_f', to the output. It uses `logical_index_vectorized` to apply the piecewise function.
}
