% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_loop_vectorized.R
\name{for_loop_vectorized}
\alias{for_loop_vectorized}
\title{Vectorized Function Using For Loop}
\usage{
for_loop_vectorized(x)
}
\arguments{
\item{x}{A numeric vector to which the `non_vectorized_function` will be applied.}
}
\value{
A numeric vector of the same length as `x`, where each element is the result of applying `non_vectorized_function` to the corresponding element of `x`.
}
\description{
Vectorized Function Using For Loop
}
\examples{
test_vector <- 1:10
result <- for_loop_vectorized(test_vector)
print(result)

}
\author{
Samy Maurer

This function applies a non-vectorized function `non_vectorized_function` to each element of a vector `x` using a for loop, effectively vectorizing the operation.
}
