% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_f.R
\name{integrate_f}
\alias{integrate_f}
\title{Computation of the integral of a Vectorized Function}
\usage{
integrate_f(a, b)
}
\arguments{
\item{a}{The lower limit of the interval over which to integrate.}

\item{b}{The upper limit of the interval over which to integrate.}
}
\value{
The computed integral value of the `logical_index_vectorized` function over the interval [a, b].
}
\description{
Computation of the integral of a Vectorized Function
}
\examples{
# Example: Compute the integral from 10 to 20
result <- integrate_f(10, 20)
print(result)
}
\author{
Samy Maurer

This function computes the integral of the `logical_index_vectorized` function over a specified interval [a, b]. It uses the `integrate` function from R.
}
