% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_vectorized_function.R
\name{non_vectorized_function}
\alias{non_vectorized_function}
\title{Non-Vectorized Piecewise Function}
\usage{
non_vectorized_function(x)
}
\arguments{
\item{x}{A numeric value for which the piecewise calculation is applied.}
}
\value{
The function returns a numeric value with those conditions:
- 0 if `x < 0`
- 15 if `0 <= x < 0.5`
- 2 if `0.5 <= x < 10`
- `8 * log(x - 9, base = 2) + 2` if `10 <= x < 15`
- `8 * log(6, base = 2) + 2` if `15 <= x <= 20`
- 0 if `x > 20`
}
\description{
Non-Vectorized Piecewise Function
}
\examples{
result1 <- non_vectorized_function(-1)
result2 <- non_vectorized_function(0.25)
result3 <- non_vectorized_function(12)
print(result1)
print(result2)
print(result3)

}
\author{
Samy Maurer

This function is a non-vectorized implementation of a piecewise function. It applies different calculations to the input `x` based on professor conditions.
}
