% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr_map_vectorized.R
\name{purrr_map_vectorized}
\alias{purrr_map_vectorized}
\title{Vectorize a Function Using `purrr::map_dbl`}
\usage{
purrr_map_vectorized(x)
}
\arguments{
\item{x}{A numeric vector to which the `non_vectorized_function` will be applied.}
}
\value{
A numeric vector where each element is the result of applying `non_vectorized_function` to the corresponding element of `x`.
}
\description{
This function vectorizes a non-vectorized function by applying it to each element of a numeric vector `x` using the `purrr::map_dbl` function.`non_vectorized_function` must be defined and takes a single numeric value as input.
}
\examples{
# Assuming non_vectorized_function is defined and takes a single numeric value
test_vector <- 1:10
result <- purrr_map_vectorized(test_vector)
print(result)

}
